Player sprite and velocity flip when camera flips
The x-velocity and sprite of the player character flip in their process functions when the "Flip" input is pressed, just like with the CameraManager. With extra time, these node scripts should be edited to rely on a function call from the CameraManager, but first the cameras must follow the player's horizontal movement.