Pauses after level flipping
The new GameManager script in the Game scene (a wrapper for the Level scene) now pauses the game briefly after flipping the level. Additionally, the Player and PlayerSprite scripts now flip the player via functions called by the Level parent, so that only CameraManager is listening for the "Flip" input and the code better follows "call down, signal up" conventions.

Jumping after flips is irritating since there's no indication that jump input will or won't register, besides the player's velocity being reapplied. A jump input buffer is needed for both this issue and for jumping in general for more enjoyment.