Adds jump buffer
Jump buffer is applied to Player script, storing jump input via the jumpBuffer float whenever the player attempts to jump during the flip pause or while in the air.

Coyote physics will be nice to make platforming more forgiving, but first death-related objects are needed to test respawning and checkpoints.