Enemies respawn after game over
Enemies are now part of an Enemy group, which reenables their visibility and their ability to kill the player after the timer runs out. This may be edited to reenable enemies after death, instead of only after a game over.

In addition, another Level scene is being created. All root Level scene nodes in the Game scene should be named "Level", as that's the name of the node the GameManager searches for at the start of each level.