Adds spikes to Level2 scene
Spikes contain a Death scene and kill the player on-contact. Level2 scene contains spike configurations that require the player to flip the level mid-jump to use their momentum for tight spaces. Spikes will need a NinePatchRect once their sprite is finalized, and this may become their root node to allow for easy scaling.